// src/types/cssClassEnums.ts
export enum AVATAR_VARIANT {
  AVATAR_ICON = 'avatar-icon',
  AVATAR_LETTER = 'avatar-letter',
  AVATAR_IMG = 'avatar-img',
  AVATAR_ICON_TEMPLATE = 'avatar-icon-template',
}

export enum AVATAR_TYPE {
  AVATAR_ROUNDED = 'avatar-rounded',
  AVATAR_CIRCULAR = 'avatar-circular',
}
// ...
export enum AVATAR_COLOR {
  AVATAR_DEFAULT_COLOR = 'avatar-default-color',
  AVATAR_RED = 'avatar-red',
  AVATAR_ORANGE = 'avatar-orange',
  AVATAR_YELLOW = 'avatar-yellow',
  AVATAR_LIME = 'avatar-lime',
  AVATAR_GREEN = 'avatar-green',
  AVATAR_TEAL = 'avatar-teal',
  AVATAR_BLUE = 'avatar-blue',
  AVATAR_INDIGO = 'avatar-indigo',
  AVATAR_PURPLE = 'avatar-purple',
  AVATAR_PINK = 'avatar-pink',
}